fix all unit tests
