+++ test/unittests/converters/adni_to_bids/data/ADNI_123_S_4567_template.xml
