+++ test/unittests/converters/adni_to_bids/data/ADNI_234_S_5678_template.xml
