+++ test/unittests/converters/adni_to_bids/data/ADNI_345_S_6789_template.xml
